+++ sources/code/Common/Input.hpp
@@ -14,7 +14,7 @@ namespace Grindstone {
 			virtual void GetMousePosition(int&, int&) = 0;
 			virtual void SetIsFocused(bool) = 0;
 			virtual bool IsFocused() = 0;
-			virtual void MouseScroll(int offsetX, int offsetY) = 0;
+			virtual void MouseScroll(float offsetX, float offsetY) = 0;
 			virtual void SetKeyPressed(Events::KeyPressCode, bool) = 0;
 			virtual bool IsKeyPressed(Events::KeyPressCode) = 0;
 			virtual void AddCharacterTyped(unsigned short character) = 0;
